refactor: move telemetry tests into separate file.
